Testsuite: building t.c using ExtUtils::CBuilder
Build.PL: add EU::CB to build_recommends
t/11-run.t: build using EU::CB

+++ t/11-run.t
@@ -1,52 +1,64 @@
+# Full test:
+# - compile t/t.c into an object file into t.o
+# - test Parse::nm->run() against t.o
+
 use strict;
 use warnings;
 use Config;
 use File::Spec;
-use Test::More tests => 9;
+use Test::More;
 use Parse::nm;
 
+BEGIN {
+    eval {
+	require ExtUtils::CBuilder;
+    };
+    if ($@) {
+	plan skip_all => 'ExtUtils::CBuilder not installed';
+    } else {
+	import ExtUtils::CBuilder;
+    }
+}
+
 my $src = File::Spec->catfile('t', 't.c');
-my $obj = "t$Config{obj_ext}";
+my $obj = eval {
+    ExtUtils::CBuilder->new(quiet => 1)->compile(source => $src);
+};
+plan skip_all => "Compile '$src' failed" if $@ || !defined $obj || !-f $obj;
 
 END {
-    if (-f $obj) {
+    if (defined $obj && -f $obj) {
         diag "Remove '$obj'";
         unlink $obj;
     }
 }
 
-diag "Compile '$src' to '$obj'...";
-# TODO Use ExtUtils::CBuilder if available
-my $exec = qx/$Config{cc} $Config{ccflags} -c $src/;
 
-SKIP: {
-    skip 'Compile failed.' => 9 unless -f $obj;
+plan tests => 8;
 
-    pass "$obj exists";
+my $count = 0;
 
-    my $count = 0;
-    Parse::nm->run(
-        files => $obj,
-        filters => [
-        {
-            name => qr/TestFunc/,
-            type => qr/[A-Z]/,
-            action => sub {
-                pass "action1 called";
-                is ++$count, 1;
-                is $_[0], "TestFunc", "arg0";
-                is $_[1], "T", "arg1";
-            }
-        },
-        {
-            name => qr/TestVar/,
-            #type => qr/[A-Z]/,
-            action => sub {
-                pass "action2 called";
-                is ++$count, 2;
-                is $_[0], "TestVar", "arg0";
-                is $_[1], 'D', 'arg1';
-            }
-        }
-    ]);
-}
+Parse::nm->run(
+    files => $obj,
+    filters => [
+    {
+	name => qr/TestFunc/,
+	type => qr/[A-Z]/,
+	action => sub {
+	    pass "action1 called";
+	    is ++$count, 1;
+	    is $_[0], "TestFunc", "arg0";
+	    is $_[1], "T", "arg1";
+	}
+    },
+    {
+	name => qr/TestVar/,
+	#type => qr/[A-Z]/,
+	action => sub {
+	    pass "action2 called";
+	    is ++$count, 2;
+	    is $_[0], "TestVar", "arg0";
+	    is $_[1], 'D', 'arg1';
+	}
+    }
+]);
